Add IP column to database.
Each mote is addressed by IPv6 address which needs to be preserved.
So adding an IP column in database.

Adding set_device_ip() so that once the device is discovered,
the IP can be easily updated.

Also made devices not be cached in a list. Since we are using sqllite,
lets query the database everytime. (Since the database is very small
all the contents should in memory all time.)